Homing seems to be working now as intended. Might need some tweaking to git rid of the 1Hz interference that was introduced with the disabling of the FIFO. Running to a non-calibrating switch now keeps the position intact. Soft limits implemented on G0 & G1 only. Removed code duplication in G161 & G162.

+++ home.h
@@ -3,7 +3,7 @@
 
 #include "bebopr.h"
 
-extern void home_axis_to_min_limit_switch( axis_e axis, int32_t* position, double feed);
-extern void home_axis_to_max_limit_switch( axis_e axis, int32_t* position, double feed);
+extern void home_axis_to_min_limit_switch( axis_e axis, int32_t* position, uint32_t feed);
+extern void home_axis_to_max_limit_switch( axis_e axis, int32_t* position, uint32_t feed);
 
 #endif	/* _HOME_H */
